feat(build-cli): New command transform:releaseNotes (#22466)
## The problem

Our goal is to automate the release process as much as possible. To that
end, now that we have a release notes generator, we'd like to
automatically upload those notes to the GitHub release as part of the
automated release.

Unfortunately, when we generate changelogs as part of release, we
consume and delete all the changesets for the release. However, we
generate a markdown file with the release notes as part of the release,
and that gets committed to the repo. Alas, we need the markdown to be
slightly different when posted to GitHub releases, and we don't have the
input changesets anymore, so we can't easily regenerate them in
automation (we'd need to find the commit in which the changesets were
consumed, go to the commit before, regenerate - it's not trivial).

## The proposed change

Since we have the release notes files in the repo, I have created a new
command, `transform:releaseNotes`, which takes the in-repo release notes
file as input, and outputs a modified markdown file for use in the
release pipeline. For example:

`flub transform releaseNotes --inFile RELEASE_NOTES/2.2.0.md --outFile
out.md`

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>
Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>