build(client): Bump build-tools dependencies to 0.46.0 (#22482)
Updates the build-tools to 0.46.0 in the client release group.

Primary changes:

The biggest changes in this release are in the type tests.

- The tests are now output in order sorted by type name.
- New in 0.46, the type tests no longer depend on any other build task,
so the task config has been updated accordingly. Also, the default
entrypoint that is tested in each package is now explicitly recorded in
package.json (those changes are in #22495).

Additional changes:

- The config now has a version property set to 1. This is a schema
version that we can change as needed when we make schema changes to the
config(s).
- The `"check:exports"` task config for test-loader-utils needed manual
updates because it does not have an api task. It has been updated with a
custom task definition.
- Several packages had task overrides that were incomplete/incorrect
because they did not inherit the global task config. Those packages have
been updated to inherit the global config.
- The same packages from the previous step also have an additional task
config for the "build:esnext" task to depend on typetests:gen; this is
needed so that ESM and CJS can be built separately. The packages in
question do not have a build:test task since they build tests as part of
their standard build. Without the build:esnext -> typetests:gen task dep
in each package, the ESM builds would not trigger typetest generation as
they should.