MergeTree: Conflict Farm Support Apply Ops During Generations (#22510)
I had thought we already had support for this as I know I prototyped it
in the past but looks like it was never committed. Regardless, this
change enables applying ops while we generate ops, which creates more
variety in the refSeqs of subsequent ops, therefore tests more cases.

However, though adding this i found some errors when run with obliterate
enable, these apply to both the current implementation, and the new
implementation in #22462. To account for this i've split the test in two
such that we expand coverage without obliterate, and keep converge for
obliterate