MergeTree: Move Obliterate to use local references (#22462)
This change was developed in a topic branch by me, @jzaffiro, and
@titrindl, as such the changes have been incrementally reviewed.

The overall change moves the implementation of merge-trees obliterate
operation from a segment-based approach to a local reference-based
approach. This motivation behind this change is to support further
changes to the obliterate operation such that it can have different
range expansion properties, or sidedness. The changes are much more
easily supported by the local reference-based approach as it allows us
to maintain and evaluate ranges independently of the physical segment
layout in the tree. Specifically, this will allow the endpoints of the
obliterate operation to exist beyond the segments that are actually
obliterated, so that the range will naturally encompass changes between
that endpoint, and the obliterated segments, which was hard to compute
in the existing implementation.

---------

Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>