Add diagnostics to ConsensusQueue in handle validation tests (#22531)
## Description

Modifies `ConsensusQueue`'s generic "store a piece of data"
implementation in `handleValidation.ts` to give more information on a
couple theoretical error methods. This test timed out on a recent
tinylicious run where based on telemetry, it seems like the test got
past the point of loading the 2nd container. Thus it was likely a
deadlock awaiting this promise. I've been unable to reproduce this
behavior locally. The goal here is to get information about the failure
mode the next time this race condition happens.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>