refactor(tree): Make properties of `FieldSchemaMetadata` readonly (#22564)
Follow-up to PR #22538. These properties were not intended to be
mutable. While this change is technically breaking, the previous PR was
made after the `2.3` release was cut, so this change is safe to make
prior to `2.4`.

Also adds a changeset for the new functionality, which was also missed
in the previous PR.