Update reconnect helper and tests for obliterate endpoint expansion  (#22575)
Some of the endpoints and expected outcomes of
`obliterate.rangeExpansion.spec.ts` had not been correct. Now that we
have a working implementation of obliterate with expansion, update these
test outcome to make sense with the implementation. Also updates the
reconnectHelper to take in additional options if necessary. Split out
from #22552, should go in after #22574