fix(ci): Fix publishing of public docker images during releases of server pipelines (#22593)
## Description

The stage that publishes the public versions of docker images for the
server components is currently broken
([example](https://dev.azure.com/fluidframework/internal/_build/results?buildId=273242&view=logs&s=2ff3421f-f27a-5211-cd8c-afeb9905f8f6),
msft internal). The problem is that `$(containerTag)` is not getting
replaced with a real value. This PR adjusts the server pipelines to
address the issue.

I'm pretty confident the problem started [when we started using the 1ES
pipeline
templates](https://github.com/microsoft/FluidFramework/pull/19382).
Before that, [the build-docker-service.yml template had variables at the
root of the
file](https://github.com/seanimam/FluidFramework/blob/f14abf74bd4474275b0ba0553728e543c13e06f3/tools/pipelines/templates/build-docker-service.yml#L100),
so all stages in that file had access to them. When we moved to 1ES
pipeline templates, those variables moved to be specified in the build
stage and the publish stages lost access to them, which results in the
`$(containerTag)` variable that the publish stage is trying to use to
not getting replaced by ADO.

Changes in this PR:
- Remove the `include-publish-docker-service.yml` template (and CI/PR
triggers off it) which is pretty redundant, only importing another file
with no special logic.
- Introduces a new template `include-vars-docker.yml` that defines the
variables which are relevant to pipelines that produce docker images.
Almost a line-by-line copy of the variables that used to be in the build
stage of `build-docker-service.yml`, with a few adjustments:
- `tagNameOnReleaseImage` was removed from the list of variables and its
logic is now inlined (and simplified) in the `publish_docker_official`
stage of `build-docker-service.yml`.
- Removed the bit that set `latestContainerTag` in the case of the
`next` branch, since we don't have a `next` branch anymore.
- The new template is now used in both the build stage and the
`publish_docker_official` (the one that fails today). No other stages
seem to need the new variables.
- Fix a small typo `1 -> i`