[main > release/client/2.1]: Handle Location redirection for getSharingInformation call (#22551) (#22621)
## Description



[AB#15188](https://dev.azure.com/fluidframework/internal/_workitems/edit/15188)

Handle location redirection when calling getSharingInformation in ODSP
Driver. Since this call does not include the v2.0 which calls the API
without the drive and item ID, these calls fail if the tenant has been
renamed. Sometimes this creates issues when creating a new page, since
the user could not get the sharing link to the new page. GetFileItem
call also returns the new siteUrl to which the file has been moved. Use
that to detect the new domain and then change the resolved url
accordingly.

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>
Co-authored-by: Frank Mueller <103605701+frankmueller-microsoft@users.noreply.github.com>