build(client): Move flub bump config settings (#22630)
The defaultInterdependencyRange setting in the fluid-build config is no
longer where settings for `flub bump` should be stored. Instead, they
should be defined in a `bump` setting in the flub config.

This change is related to #22628. This change will make the changes in
#22628 irrelevant in main and future release branches, but they are
still needed so the tools work against release branches without this
change (2.3 client and before).