fix(build-cli): Load interdependency ranges from fluid-build config for back-compat (#22628)
## Summary

`flub bump` can be configured to use different default settings for
different release groups. Those settings are in the root flub config -
which historically has been the same as the fluid-build config file.
This change updates the code to add back-compat handling. If the
settings in the flub config are not found, the fluid-build config will
be checked. This back-compat code can be removed as soon as we no longer
support any release branches without the new config.

This change does unfortunately add back a dependency between build-tools
and version-tools. Fortunately, that can be removed once we no longer
need to support the older configs.

#22630 removes the need for this change by updating the config - but
this change is still needed for back-compat with earlier configs.

## Background

When the fluid-build and flub configs were split in #21967, it didn't
take into account the config settings in the fluid-build config that
control the default interdependency range for each release group. Those
settings are intended to be in the `bump` section of the flub config,
but the fluidBuild config file in the repo doesn't include those
settings, and there was no back-compat code that loaded them as far as I
can tell.

In #21967, though, the settings were removed from the old fluid-build
config types (see the removal of lines 330-335 in fluidRepo.ts:
[change](https://github.com/microsoft/FluidFramework/commit/f667535a3f4f893115c3360a78d07acfa9d24456?diff=split&w=0#diff-9a0994a9d2ddb86f6f1e53fef4e8de22cf87ced6dd81bd0e8866784c2679f450L330)

That was an unintended breaking change because the old settings are not
loaded when using new versions of flub. This change adds the missing
back-compat code to load those settings.