Generate revertibles for commits incoming from a merge (#22644)
## Description

This corrects some behavior in the (currently alpha) branching APIs.
Currently, merging one or more commits into a target branch from a
source branch does not generate revertibles on the target branch. This
PR updates it so that each merge commit fires a "commitApplied" event
and generates the proper revertible.