feat(api-markdown-documenter): Improve HTML transform API and promote to public (#22652)
Allow users to invoke `toHtml` functions with either a partial config or
a full transform context. Also promotes `toHtml` APIs to `@public` so we
can use them in the website build.

Allows updating custom table renderer in `fluidframework.com`'s API docs
build to leverage the transformation and
[hast-util-to-html](https://www.npmjs.com/package/hast-util-to-html) to
simplify the logic as follows:


![image](https://github.com/user-attachments/assets/bfcfb23e-950c-42a8-bbe9-9b3c85bdd428)


[AB#8866](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8866)

The above bug seems to stem from our "pretty format" HTML rendering. A
cursory investigation didn't yield useful context on exactly what was
going wrong in that code. But the idea has been for quite a while to
eventually remove that code and replace it with the HAST-based `toHtml`
transformation logic anyways. I have confirmed that switching to this,
and using an existing open-source `hast-util-to-html` HAST renderer
resolves the issue.