Remove dead ordered client election code (#22488)
[AB#15192](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/15192)

Removes dead code from `OrderedClientElection`. 

Historical note: we had developed a feature that elected a new
summarizer client when the current summarizer client was failing to
summarize. When we shipped the feature and turned it on, we ran into
many buggy issues with multiple summarizers. Since then, we've turned
off the feature and have some dead code. This PR aims to remove that
dead code.