fix: Correct text formatting handling in `toHtml` transformation (#22672)
The Documentation domain supports hierarchical text formatting
overrides. This includes setting formatting at a position in the tree,
and disabling it at a lower level in the tree.

HTML does not support this kind of hierarchical format disabling. The
way we were transforming Documentation domain input to HTML ignored
this, and was fundamentally broken. This change fixes the issue by
tracking text formatting context while transforming the tree, and
injecting formatting tags only at the lowest levels.

See updated tests for examples of the changes.