improvement(fluid-build): Pass BuildContext object to tasks (#22666)
Tasks in fluid-build often need to know what the root of the repo is and
have access to the fluidBuild config. Historically, individual tasks
have stored these values themselves where needed. This change adds a new
`BuildContext` object that is available to all Task subclasses.

```ts
/**
 * A context object that is passed to fluid-build tasks. It is used to provide easy access to commonly-needed metadata
 * or tools.
 */
export interface BuildContext {
	/**
	 * The fluid-build configuration for the repo.
	 */
	fluidBuildConfig: IFluidBuildConfig | undefined;

	/**
	 * The absolute path to the root of the repo.
	 */
	repoRoot: string;

	/**
	 * A GitRepo object that can be used to call git operations.
	 */
	gitRepo: GitRepo;
}
```

There was an existing BuildContext class that was used only in the
BuildPackage class. I renamed it to BuildGraphContext and it implements
the BuildContext interface.

I renamed the properties on BuildGraph and BuildPackage to the generic
`context` since it now contains everything from both the more widely
used BuildContext and the BuildGraphContext.

Related to
[AB#13203](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13203)
and
[AB#13202](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13202)