refactor(api-markdown-documenter): Update HTML rendering functionality to leverage HAST domain and existing HAST rendering libraries (#22664)
Removes A LOT of bespoke HTML rendering logic, replacing it with
[hast-util-to-html](https://www.npmjs.com/package/hast-util-to-html) and
[hast-util-format](https://www.npmjs.com/package/hast-util-format) to do
the rendering on our behalf.

Note that there are A LOT of test asset diffs here. The rendering
libraries mentioned above make different styling choices than we made in
our bespoke rendering. In most cases these changes are trivial, or
arguably readability improvements. Regardless, this PR makes the
formatting no longer a direct responsibility of this library, which is a
win IMO.