(tree): Add explicit revision tag to all changes (#22525)
## Description
When the default edit builder generates a change, it does not have a
revision tag associated with it. The revision tag is added later by
calling `ChangeRebaser.changeRevision`. This PR updates these changes to
have revision tag associated with it.
It also updates the `invert` function to generate changes with revision
tags.


[AB#13508](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13508)

---------

Co-authored-by: Yann Achard <achardy@microsoft.com>