Added support for health probes for all HTTP services in Routerlicious (#22635)
## Description

This PR adds support for the following endpoints for `Riddler, Nexus,
and Alfred`:

1) `/healthz/startup`: Startup readiness check endpoint
2) `/healthz/ready`: Service lifecycle readiness check endpoint
3) `/healthz/ping`: Liveness endpoint. This endpoint was not added for
`Alfred` as it already has an existing ping endpoint `/api/v1/ping`

These are needed to support Kubernetes Health Checks.

The startup endpoint relies on a new singleton class introduced in this
PR - `StartupChecker`. This class returns the `startup` status as
`isReady: true` after the service runner is created.

The readiness endpoint would need a custom implementation of
`IReadinessCheck`. If this is not provided, the endpoint will not be
created.

To support HTTP endpoints in Nexus, it also adds a request listener to
the HTTP server setup in Nexus.

## Breaking Changes

Adds customizations to the ResourceFactory and Runners each of the
service mentioned above. These are used to inject an implementation of
`IReadinessCheck`.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>