Added support for health probes to Gitrest and Historian (#22710)
## Description

This PR takes in the r11 changes -
https://github.com/microsoft/FluidFramework/pull/22635 - and adds
support for the `/healthz` endpoints for `Historian` and `Gitrest`.

1. `/healthz/startup`: Startup readiness check endpoint
2. `/healthz/ready`: Service lifecycle readiness check endpoint
3. `/healthz/ping`: Liveness endpoint. This endpoint was not added for
`Historian` as it already has an existing ping endpoint `/repos/ping`
4. 
These are needed to support Kubernetes Health Checks.

The readiness endpoint would need a custom implementation of
IReadinessCheck. If this is not provided, the endpoint will not be
created.

## Breaking Changes
Adds customizations to the ResourceFactory and Runners each of the
service mentioned above. These are used to inject an implementation of
IReadinessCheck.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>