Fix applies stashed ops with no saved ops test (#22712)
In odsp, when you start a container in a paused state by using
deltaConnection: none load mode, it does not load from the most recent
summary (not sure if it is intermittent or not). For the purposes of e2e
test "applies stashed ops with no saved ops" the base snapshot that is
loading from does not necessarily needs to be the most recent one but
only that we don't have save ops, which we are asserting correctly, so
removing the unnecessary assert is making the test to consistently pass.