MergeTree: Seperate segment visabilty interfaces (#22693)
This change is a precursor to removing deprecations from ISegment:
`segmentGroups`, `propertyManager` and `localRefs`

This change expands on the distinction between `ISegment` and
`ISegmentLeaf` by introducing `ISegmentInternal`. This leads to a
hierarchy of interfaces tired by their exposure:

- Legacy: `ISegment`
- Internal: `ISegmentInternal`
- Unexported:  `ISegmentLeaf`

This allows different exposure layers to work with different
capabilities and will let us hide more the of api that isn't meant to be
used by customers.