fix(client-presence): correct secondary join response (#22673)
When a ClientJoin message is received, clients that are not listed as
primary nominees to respond should order themselves in some way (to
avoid a large number of responses where O(n) = n^2). For clients that
are quorum members, the intended design is to use relative quorum
ordering.
The prior implementation used quorum member sequenceNumber as a simple
ordered index within the members. However, that is not what
sequenceNumber represents. It is an op related number and may be quite
high. That could result in an extremely delayed response.
sequenceNumber can be used to generated relative order among quorum
members. Update the implementation to do just that. (Ordering for
non-quorum members is unchanged.)