refactor: Update React dependencies (#22726)
## Description

Update React from 17 to 18 in our examples and in Devtools. Also cleans
up the `@types/react-dom` dev dependency from packages that don't import
`react-dom`, and adds `@types/react` and `@types/react-dom` as dev
dependencies to a package that uses both of them but didn't have a
dependency on the types.

This is motivated by work on the Fluid AI Collab experiment which uses a
NextJS app as an example (because it wants to run some code server side)
and NextJS requires React 18.