Use diff instead of status for policy check done file (#22597)
## Description

Policy check should probably be invalidated when files are deleted.

`getModifiedFiles` contained some (unstaged) but not all (staged)
deleted files. From its description, its unclear what exactly it should
include, but its only used by FlubCheckPolicyTask which behaves
incorrectly if there are deleted files, regardless of if
`getModifiedFiles` includes them or not (crashes if they are included,
has an incremental build issue if not).

This change avoids it processing a list of modified files entirely, and
just hashes the result of `git diff HEAD` and the head revision, which
is a simple git powered summary of the current working copy state.

I also added some comments since I ran into an entire class inheritance
chain with no doc comments and method overload chains with no doc
comments, which made this code hard to understand. I'm not confident
these comments are correct so they should be reviewed carefully.

This is inspired by
https://github.com/microsoft/FluidFramework/pull/22586 which made me
struggle to determine if `getModifiedFiles` is correct and made me want
to delete it or clarify its specification.