refactor(build-cli): Move tests alongside source (#22684)
Moves the tests for build-cli inside the source folder, so that it
matches the patterns used elsewhere in the repo and release group. Also
updates the tsconfigs to use the shared config in build-common.

Required several lint disables in the tests, but they were all
reasonable IMO. If needed those can be cleaned up in future changes.