Fix offline blob tests (#22744)
in ODSP, we can't load a container offline and send a server request to
create a blob because our call would not be complete, specifically we
would lack the epoch value which is set until we connect to the delta
stream, ending up with a 'ODSP fetch error [400]' response.

Even though this problem is worth tackling separately I believe both
affected tests are not specifically testing this behavior so I changed
them slightly to send a createBlob request while being disconnected
instead of using our loadOffline function. Both tests are connecting
anyway after creating the blob and the relevant asserts happen after
that part of those tests.

I added another test which replicates the issue and skip it in ODSP for
now. If we want to solve the initial issue, we would need to stash the
epoch info in our serialized state.
[AB#19035](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/19035)

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>