Remove READ support for compatDetails on ContainerRuntime ops (#22715)
#### Description 
Search for compatDetails in the code - all some usages can now be
removed. Maintaining this behavior has been a pain in tests, and it is
not a good feature.

It was only ever used for GC, and wasn't as useful (or safe) as
expected. See discussion in ADO:
https://dev.azure.com/fluidframework/internal/_workitems/edit/17024 -
RC1 clients don't know about TombstoneLoaded subtype

Fixes issue:
[AB#17024](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/17024)

The full removal of this feature is planned for next sprint (10/14 -
11/1)

---------

Co-authored-by: Mark Fields <markfields@users.noreply.github.com>