Offline: Properly handle ID Allocation batches (#22739)
ID Allocation ops are not resubmitted one-by-one like other ops, but
rather the ID Compressor is told to generate new ops once and for all
right at the start of the Resubmit flow (before calling
`PendingStateManager.replayPendingStates`). The original Batch ID isn't
available at that point, so there's no way to correlate the new ID
Allocation Batch to the original one.

In a forked container scenario, we can ignore the Batch IDs of ID
Allocation batches, knowing that the main batch will also be duplicated
and we'll detect that.