(tree) Added a beta API to deep clone tree nodes (#22719)
## Description
Added a beta API to deep clone nodes in a tree. The properties of this
API are:
- It deep clones all persisted data associated with the node and its
subtree.
- It can clone primitive types as well.
- It can clone both hydrated and unhydrated nodes.

## Future considerations
One useful property of this API could be to replace identifiers in the
node's subtree. However, I am choosing to not do this in the first
iteration of this API. We can consider doing this as a follow up or when
there is a customer ask which will give us more details into what the
requirements are and how best to design it.


[AB#13000](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13000)