Fix 0x11f raise condition (#22780)
While stashing blobs, we need to first send the blob attach op before
resolving the blob handle to keep order of ops correctly (blob attach op
and op referencing the blob). This change of order causes the stashed
container to have a blob attach op whose blob won't be used in case of
aborting the closeAndGetPendingLocalState call.

Instead of cleaning up the pending state, for now we will just accept
the fact that the blob attach op will be stashed with no usage. We could
come up with an optimization of this later on when our partners catch up
with our latest offline features.