fix(generate:changelog): Calculate correct changeset version (#22796)
We run `changeset version` during changelog generation, which causes
changesets to be consumed and the package.json versions to be bumped. We
need to calculate the same "target version" that the changesets tools
calculate so we can correctly replace the output of the changelogs to
the correct versions and do our other cleanup.

This change fixes the calculation to take into account the bump ranges
in the changesets.