refactor: Disable CodeQL in test pipelines (#22802)
## Description

Disables the auto-injection of CodeQL tasks in our test pipelines. The
CodeQL tasks always "fail" (the ADO step still succeeds but the task
can't find source code so it can't run CodeQL and logs warnings) in the
test pipelines for which we don't checkout the repo, which is the case
for our test pipelines.
[Example](https://dev.azure.com/fluidframework/internal/_build/results?buildId=300062&view=logs&j=26483432-40f5-552c-5792-70aed8cde738&t=fb288d9b-a668-57e8-da86-3cec1930f710&l=55)
(msft internal). The task itself says that it should only run for build
pipelines (where source code exists).